Fix scene GeometryTest FIx UI Hitpoints
