+++ Assets/_Karl/Scripts/App/App.cs
@@ -1,3 +1,4 @@
+using Engine3D;
 using UnityEngine;
 
 public class App : Singleton<App>
@@ -15,5 +16,26 @@ public class App : Singleton<App>
 #else
         Application.Quit();
 #endif
+    }
+
+
+    protected override void Awake()
+    {
+        base.Awake();
+        
+        GameResources.Initialize();
+        AssetManager assetManager = AssetManager.Singelton; // force initialization
+        GameState.AudioSystem = new Audio.AudioSystem();
+        GameState.AudioSystem.SetAudioSource(GetComponent<AudioSource>());
+        
+        GameState.TileSpriteAtlasManager.UpdateAtlasTextures();
+        GameState.SpriteAtlasManager.UpdateAtlasTextures();
+        GameState.IsInitialized = true;
+    }
+    
+    protected override void OnDestroy()
+    {
+        base.OnDestroy();
+        GameState.IsInitialized = false;
     }
 }
