+++ Assets/_Karl/Scripts/Player/PlayerInputStandalone.cs
@@ -45,61 +45,59 @@ public class PlayerInputStandalone : BaseMonoBehaviour
 
     private void UpdateInputs()
     {
-        if(!App.Instance.GetPlayer().GetRenderer().GetAgent().agentStagger.Stagger)
-        {
-            // Main menu
-            if (IsKeyDown(_mainMenu)) _input.DoOpenEscapeMenu();
+        // Main menu
+        if (IsKeyDown(_mainMenu)) _input.DoOpenEscapeMenu();
 
-            if (IsKeyDown(_inventory)) _input.DoToggleInventory();
+        if (IsKeyDown(_inventory)) _input.DoToggleInventory();
 
-            // Movement
-            bool left = IsKey(_moveLeft);
-            bool right = IsKey(_moveRight);
-            bool up = IsKey(_moveUp);
-            bool down = IsKey(_moveDown);
+        // Movement
+        bool left = IsKey(_moveLeft);
+        bool right = IsKey(_moveRight);
+        bool up = IsKey(_moveUp);
+        bool down = IsKey(_moveDown);
+    
+        // Look Target
+        _input.DoPlayerLookTarget(_inputArea.GetLastMove().position);
+
+        // Crouch
+        if (IsKeyDown(_crouch)) _input.DoPlayerCrouchBegin(left, right);
+        if (IsKeyUp(_crouch)) _input.DoPlayerCrouchEnd(left, right);
+
+        // Walk
+        _input.DoPlayerWalk(left, right);
+    
+        // Sprint
+        if (IsKey(_sprint)) _input.DoPlayerSprint(left, right);
+
+        // Dash
+        if(IsKeyDown(_speedDash)) _input.DoPlayerDash(left, right);
+    
+        // Jump
+        if(IsKeyDown(_jump)) _input.DoPlayerJump();
+    
+        // Jetpack
+        if(IsKeyDown(_jetpack)) _input.DoPlayerJetpackBegin();
+        if(IsKeyUp(_jetpack)) _input.DoPlayerJetpackEnd();
+    
+        // Fire
+        if(IsKeyDown(_fire)) _input.DoPlayerFire();
+
+        // Second Action
+        if (IsKeyDown(_secondAction)) _input.DoSecondAction();
+
+        // Reload
+        if (IsKeyDown(_reload)) _input.DoPlayerReload();
+
+        // Drop item
+        if (IsKeyDown(_dropItem)) _input.DoPlayerDropItem();
+    
+        // Take Screenshot
+        if (IsKeyDown(_screenShot)) _input.DoScreenshot();
+
+
+        // Footer quick slots inventory
+        UpdateQuickSlots();
         
-            // Look Target
-            _input.DoPlayerLookTarget(_inputArea.GetLastMove().position);
-
-            // Crouch
-            if (IsKeyDown(_crouch)) _input.DoPlayerCrouchBegin(left, right);
-            if (IsKeyUp(_crouch)) _input.DoPlayerCrouchEnd(left, right);
-
-            // Walk
-            _input.DoPlayerWalk(left, right);
-        
-            // Sprint
-            if (IsKey(_sprint)) _input.DoPlayerSprint(left, right);
-
-            // Dash
-            if(IsKeyDown(_speedDash)) _input.DoPlayerDash(left, right);
-        
-            // Jump
-            if(IsKeyDown(_jump)) _input.DoPlayerJump();
-        
-            // Jetpack
-            if(IsKeyDown(_jetpack)) _input.DoPlayerJetpackBegin();
-            if(IsKeyUp(_jetpack)) _input.DoPlayerJetpackEnd();
-        
-            // Fire
-            if(IsKeyDown(_fire)) _input.DoPlayerFire();
-
-            // Second Action
-            if (IsKeyDown(_secondAction)) _input.DoSecondAction();
-
-            // Reload
-            if (IsKeyDown(_reload)) _input.DoPlayerReload();
-
-            // Drop item
-            if (IsKeyDown(_dropItem)) _input.DoPlayerDropItem();
-        
-            // Take Screenshot
-            if (IsKeyDown(_screenShot)) _input.DoScreenshot();
-
-
-            // Footer quick slots inventory
-            UpdateQuickSlots();
-        }
     }
 
     private void UpdatePlayer()
