+++ Assets/_Karl/Scripts/Renderer/AgentRenderer.cs
@@ -68,8 +68,7 @@ public class AgentRenderer : BaseMonoBehaviour
         if (_agent.hasAgentStats)
         {
             _hitpoints = Instantiate(AssetManager.Singelton.GetPrefabUIHitpoints(), _pivotHead, false);
-            _hitpoints.SetHitpoints(new ContainerInt(_agent.agentStats.Health.GetValue(), _agent.agentStats.Health.GetMax(),
-                _agent.agentStats.Health.GetMin()));
+            _hitpoints.SetHitpoints(_agent.agentStats.Health);
 
             bool factionFriendly = false;
             if(!agent.isAgentPlayer){
